Move Record model in to CollectionModel.ts, start implementing the collection/record model utilities
